[SDK] Require DType input to ops.range
By explicitly setting a DType it's much easier to infer user intent for
the output of `range`. The Graph API tries to be simple, predictable,
and explicit, and explicit DTypes aid greatly in this.

As a result the logic for ranges was simplified greatly, and many cases
which previously required explicitly creating a constant or specifying
an output dimension no longer do.

MAX_INTEGRATION_TESTS_REV_ID: 0071dd7142b1d450f82273ddedeb7cc45600483c

+++ max/tests/integration/API/python/tensor/test_functional_other.py
@@ -241,7 +241,7 @@ def test_permute():
 
 def test_range():
     device_ref = DeviceRef.from_device(DEVICE)
-    result = F.range(0, 10, 1, device=device_ref)
+    result = F.range(0, 10, 1, dtype=DType.int32, device=device_ref)
     result._sync_realize()
     assert result.real
